now the pip will install stuff based on whether the requirements.txt exists

+++ {{cookiecutter.project}}/bin/vEnv.sh
@@ -1,5 +1,16 @@
 #!/bin/bash
 
+#----------------------------------------------
+# Note that this is the standard way of doing 
+# things in Python 3.6. Earlier versions used
+# virtulalenv. It is best to convert to 3.6
+# before you do anything else. 
+# Note that the default Python version in 
+# the AWS Ubuntu is 3.5 at the moment. You
+# will need to upgrade the the new version 
+# if you wish to use this environment in 
+# AWS
+#----------------------------------------------
 python3 -m venv env
 
 # this is for bash. Activate
@@ -8,28 +19,37 @@ python3 -m venv env
 source env/bin/activate 
 
 pip3 install --upgrade pip
-pip3 install pytest
-pip3 install pytest-cov
-pip3 install sphinx
 
-# Utilities
-pip3 install ipython
-pip3 install tqdm
+if [ -e requirements.txt ]; then
 
-# scientific libraries
-pip3 install numpy
-pip3 install scipy
-pip3 install pandas
-pip3 install sklearn
+    pip3 install -r requirements.txt
 
-# database stuff
-pip3 install psycopg2
+else
 
-# Charting libraries
-pip3 install matplotlib
-pip3 install seaborn
-pip3 install folium
+    pip3 install pytest
+    pip3 install pytest-cov
+    pip3 install sphinx
 
-pip3 freeze > requirements.txt
+    # Utilities
+    pip3 install ipython
+    pip3 install tqdm
+
+    # scientific libraries
+    pip3 install numpy
+    pip3 install scipy
+    pip3 install pandas
+    pip3 install sklearn
+
+    # database stuff
+    pip3 install psycopg2
+
+    # Charting libraries
+    pip3 install matplotlib
+    pip3 install seaborn
+    pip3 install folium
+
+    pip3 freeze > requirements.txt
+
+fi
 
 deactivate
